+++ src/main/java/net/atos/entng/support/GridFSAttachment.java
@@ -2,22 +2,22 @@ package net.atos.entng.support;
 
 public class GridFSAttachment extends Attachment
 {
-    public GridFSAttachment(Integer bugTrackerId, String name)
+    public GridFSAttachment(Long bugTrackerId, String name)
     {
         super(bugTrackerId, name);
     }
 
-    public GridFSAttachment(Integer bugTrackerId, String name, Integer size)
+    public GridFSAttachment(Long bugTrackerId, String name, Integer size)
     {
         super(bugTrackerId, name, size);
     }
 
-    public GridFSAttachment(Integer bugTrackerId, String name, String fileSystemId)
+    public GridFSAttachment(Long bugTrackerId, String name, String fileSystemId)
     {
         super(bugTrackerId, name, null, fileSystemId);
     }
 
-    public GridFSAttachment(Integer bugTrackerId, String name, Integer size, String fileSystemId)
+    public GridFSAttachment(Long bugTrackerId, String name, Integer size, String fileSystemId)
     {
         super(bugTrackerId, name, size, null, fileSystemId);
     }
